Change the image size in posts when changing layouts
Since the styling of images in posts in the list layout were getting set on the image itself, I needed to add logic to the PostView viewmodel to expose an attribute that sets which image size to display for a given layout. The template then uses this attribute to determine how the <img> tag should be configured.