Ensure that object tags in buttons are clickable
An anchor inside of a button tag was not rendering in IE9. This change moves the object-btn tags outside of their buttons to be a sibling element. This ensures that the anchors are clickable in IE and workaround the issue where embedded objects are not clickable in IE when inside a button.